chore: get corejs version
