+++ packages/core/src/plugins/swc.ts
@@ -1,3 +1,4 @@
+import fs from 'node:fs';
 import path from 'node:path';
 import {
   type Polyfill,
@@ -5,7 +6,6 @@ import {
   applyScriptCondition,
   cloneDeep,
   deepmerge,
-  getCoreJsVersion,
 } from '@rsbuild/shared';
 import type { SwcLoaderOptions } from '@rspack/core';
 import { PLUGIN_SWC_NAME } from '../constants';
@@ -136,6 +136,17 @@ export const pluginSwc = (): RsbuildPlugin => ({
     });
   },
 });
+
+const getCoreJsVersion = (corejsPkgPath: string) => {
+  try {
+    const rawJson = fs.readFileSync(corejsPkgPath, 'utf-8');
+    const { version } = JSON.parse(rawJson);
+    const [major, minor] = version.split('.');
+    return `${major}.${minor}`;
+  } catch (err) {
+    return '3';
+  }
+};
 
 async function applyCoreJs(
   swcConfig: SwcLoaderOptions,
